@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import 'rxjs/add/operator/map';
 
 @Injectable()
@@ -12,11 +12,11 @@ export class ImageService {
   private URL: any = this.API_URL + this.API_KEY + '&q='; // join all variables
   private perPage = '&per_page=200'; // Nº od image what to show per page
 
-  constructor( private http: Http ) { }
+  constructor( private http: HttpClient ) { }
 
   getImage(query) {
     return this.http.get(this.URL + query + this.perPage)
-    .map(response => response.json());
+    .map(response => response)
   }
 
 }
